requirements.txt
fastapi~=0.115.8
uvicorn~=0.34.0
python-dotenv~=1.0.1
pyodbc~=5.2.0
jinja2
